+++ src/lib/components/UI.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import Zoom from '$lib/components/Tools/Zoom.svelte';
 	import ServerId from './ServerId.svelte';
-	import Menu from './Menu.svelte';
+	import Menu from './Toolbar.svelte';
 	import ChatButton from './Chat/ChatButton.svelte';
 	import Navbar from './Navbar/Navbar.svelte';
 	import Chat from './Chat/Chat.svelte';
